Extracts Message component from OAuthCallback page
