+++ waspc/e2e-test/test-outputs/waspCompile-golden/files.manifest
@@ -207,6 +207,7 @@ waspCompile/.wasp/out/web-app/public/.gitkeep
 waspCompile/.wasp/out/web-app/public/favicon.ico
 waspCompile/.wasp/out/web-app/public/manifest.json
 waspCompile/.wasp/out/web-app/scripts/validate-env.mjs
+waspCompile/.wasp/out/web-app/src/components/Message.tsx
 waspCompile/.wasp/out/web-app/src/index.tsx
 waspCompile/.wasp/out/web-app/src/logo.png
 waspCompile/.wasp/out/web-app/src/router.tsx
