+++ waspc/e2e-test/test-outputs/waspComplexTest-golden/files.manifest
@@ -464,6 +464,7 @@ waspComplexTest/.wasp/out/web-app/public/manifest.json
 waspComplexTest/.wasp/out/web-app/scripts/validate-env.mjs
 waspComplexTest/.wasp/out/web-app/src/auth/pages/OAuthCallback.tsx
 waspComplexTest/.wasp/out/web-app/src/auth/pages/createAuthRequiredPage.jsx
+waspComplexTest/.wasp/out/web-app/src/components/Message.tsx
 waspComplexTest/.wasp/out/web-app/src/index.tsx
 waspComplexTest/.wasp/out/web-app/src/logo.png
 waspComplexTest/.wasp/out/web-app/src/router.tsx
