+++ waspc/e2e-test/test-outputs/waspJob-golden/files.manifest
@@ -242,6 +242,7 @@ waspJob/.wasp/out/web-app/public/.gitkeep
 waspJob/.wasp/out/web-app/public/favicon.ico
 waspJob/.wasp/out/web-app/public/manifest.json
 waspJob/.wasp/out/web-app/scripts/validate-env.mjs
+waspJob/.wasp/out/web-app/src/components/Message.tsx
 waspJob/.wasp/out/web-app/src/index.tsx
 waspJob/.wasp/out/web-app/src/logo.png
 waspJob/.wasp/out/web-app/src/router.tsx
