+++ waspc/e2e-test/test-outputs/waspMigrate-golden/files.manifest
@@ -211,6 +211,7 @@ waspMigrate/.wasp/out/web-app/public/.gitkeep
 waspMigrate/.wasp/out/web-app/public/favicon.ico
 waspMigrate/.wasp/out/web-app/public/manifest.json
 waspMigrate/.wasp/out/web-app/scripts/validate-env.mjs
+waspMigrate/.wasp/out/web-app/src/components/Message.tsx
 waspMigrate/.wasp/out/web-app/src/index.tsx
 waspMigrate/.wasp/out/web-app/src/logo.png
 waspMigrate/.wasp/out/web-app/src/router.tsx
